Hardened the User Info Cache code so it would not error if the data somehow was cleared to Null and written to the credentials storage. It will also write an empty dict to the credentials instead of a dict of nulls.
Stripped the leading and trailing whitespace characters during 2FA code inputs so the whitespace wouldn't make your 2FA fail.

Properly type-hint 2FA code so it would resolve nicely in any IDE or MyPy

Fixed broken urls in current_version,json
Updated version.py to have 1.0.7 with relevant changelog information

Hardened the User Info Cache code so it would not error if the data somehow was cleared to Null and written to the credentials storage. It will also write an empty dict to the credentials instead of a dict of nulls.

Stripped the leading and trailing whitespace characters during 2FA code inputs so the whitespace wouldn't make your 2FA fail.

Properly type-hint 2FA code so it would resolve nicely in any IDE or MyPy

Fixed broken urls in current_version,json

update gitignore to avoid .gz files. The mac with installer is uses that compression. partially updated the batch file to deal with issues, still needs some tweaking

update auto-copy-windows to use full path to the zip. Not sure why necessary but it's not a difficult fix.