Implemented MVC authentication in Model and View
Implemented view. implemented auth code in controller, and skeleton in the model.

moved the user info cache outside of cache list, then simplified and renamed it. It is no longer going to persist in any manner as a security enhancement.

enums.py is now obsolete and removed.
display page was an attempt to bridge display helper, that is now taken care of in MVC Classes as a Class Enum.
Cache Handler is prototyped in the caches class now so it is redundant and therefore removed.
SteamAuthPollingData similarly is now handled within MVC Classes.

fixed a bad type hint in websocket client... but it'll be removed so this is pointless, lol.